requirements.txt
python-chess
cairosvg
Pillow